Define DB fixtures in a shared place
